Attach graph node index on PGraph metadata

+++ include/gengodb/runtime/BuiltinGraphs.h
@@ -186,7 +186,7 @@ public:
     
     struct Metadata {
         inline void set_name_mapping(std::function<std::string(int32_t)> names) { names_ = names; }
-        inline void set_id_mapping(std::function<uint64_t(int32_t)> uid, std::function<uint64_t(int32_t)> local) { uid_ = uid; local_ = local; }
+        inline void set_id_mapping(std::function<uint64_t(int32_t)> uid, std::function<int32_t(uint64_t)> local) { uid_ = uid; local_ = local; }
         inline std::string get_node_name(int32_t id) const { return names_(id); }
         inline uint64_t uid(int32_t local_id) const { return uid_(local_id); }
         inline uint64_t local_id(uint64_t uid) const { return local_(uid); }
